requirements.txt
# Image Captioning Project Dependencies
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.8.0

# Data Processing & Utilities
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars & Logging
tqdm>=4.65.0
tensorboard>=2.13.0

# JSON & File Handling
requests>=2.31.0
urllib3>=1.26.0

# COCO Dataset API (for dataset downloading)
pycocotools>=2.0.6

# Jupyter Notebook Support (for Colab)
jupyter>=1.0.0
ipython>=8.14.0

# Model Evaluation Metrics
scikit-learn>=1.3.0
nltk>=3.8.0

# Web App Dependencies (for later deployment)
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0

# Additional Utilities
python-dotenv>=1.0.0
pathlib2>=2.3.7

Optional: For better performance
numba>=0.57.0  
Development Tools (Optional)
pytest>=7.4.0  # For testing
black>=23.0.0   # Code formatting
flake8>=6.0.0   # Code linting